shared library support
Adds support for shared libraries. Specifically, Tockloader is
responsible for resolving references to shared library functions
that the application makes. Tockloader can read all of the board's flash
memory, so it knows where the shared library binary exists in flash.
It also has information about where the function exists within
the shared library binary thanks to the relocation information passed
from elf2tab (via a TOML file in the TAB tar).

To support calling shared libraries via function pointers, Tockloader
sets aside a bit of flash memory to store the actual address of the
shared library function code. This way, there can be a level of an
indirection when the app code attempts to call a shared library
function.